pyproject.toml
[project]
name = "db-perf"
version = "0.1.0"
description = ""
authors = [
    {name = "David Onuh",email = "davidonuh1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "db_perf", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
